+++ server/aesdsocket.c
@@ -1,32 +1,43 @@
+/**
+ * @file    aesdsocket.c
+ * @brief   This program creates a socket server that listens on port 9000, accepts incoming connections,
+ *          and receives data from clients. It stores the received data in a file at "/var/tmp/aesdsocketdata"
+ *          and sends back the accumulated data to the clients.
+ *			
+ * @date    October 5, 2023
+ * @author  Rishikesh Sundaragiri
+ */
+#include <stdio.h>
+#include <stdbool.h>
+#include <string.h>
 #include <sys/types.h>
 #include <sys/socket.h>
-#include <netdb.h>
+#include <netdb.h> //has the struct addrinfo variable here
 #include <syslog.h>
 #include <unistd.h>
 #include <signal.h>
 #include <sys/stat.h>
 #include <stdlib.h>
-#include <stdio.h>
 #include <fcntl.h>
 #include <sys/wait.h>
-#include <string.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
 #include <getopt.h>
 #include <errno.h>
 
-
 #define PORT "9000"
 #define BACKLOG 10
-#define CHUNK_SIZE 400
-#define TEST_FILE  "/var/tmp/aesdsocketdata"
+#define ERROR_CODE -1
+#define DATA_FILE "/var/tmp/aesdsocketdata"
 
-int serv_sock_fd,client_sock_fd,output_file_fd, total_length,len,capacity,counter=1,close_err;
-struct sockaddr_in conn_addr;
-char IP_addr[INET6_ADDRSTRLEN];
-		
 
-//Below function referenced from https://beej.us/guide/bgnet/html/
+int main_sockfd = 0;
+int client_sockfd = 0;
+int data_file_fd = 0;
+
+/*
+ * Reference : https://beej.us/guide/bgnet/html/#cb47-58
+*/
 void *get_in_addr(struct sockaddr *sa)
 {
     if (sa->sa_family == AF_INET) {
@@ -36,330 +47,331 @@ void *get_in_addr(struct sockaddr *sa)
     return &(((struct sockaddr_in6*)sa)->sin6_addr);
 }
 
-//Function handles all open files when there is an error
-void close_all()
+/*
+ *  Executes whenever a  SIGINT or SIGTERM is received
+ *  parameter: int sig is the received signal number
+ *  NOTE : My initial idea was to keep the signal handler as short as possible. So I just set a flag here and used to ckeck that flag inside while(1)
+ *  to detech a signal occurance. But by doing that the Crtl+C command was not working. So I had to put the entire logic of cleanup inside the signal handler 
+ *  to make it work. But I still hesitate to put the cleanup code here because its not safe.
+ */
+void sig_handler(int sig)
 {
-	//Functions called due to an error, hence close files errno not 
-	//checked in this function
-	//All close errors handled in signal handler when no error occured
-	close(client_sock_fd);
-	//Close server socket fd
-	close(serv_sock_fd);
-	//Close regular file - output file fd
-	close(output_file_fd);
-	
-	//After completing above procedure successfuly, exit logged
-	syslog(LOG_DEBUG,"Caught signal, exiting");
-	
-	//Close the syslog once all of logging is complete
-	closelog();
-	
-	//Delete and unlink the file
-	remove(TEST_FILE);
+    if(-1 == close(main_sockfd))
+    {
+        perror("close(main_sockfd)");
+        exit(ERROR_CODE);  
+    }
+    if(-1 == close(data_file_fd))
+    {
+        perror("close(data_file_fd)");
+        exit(ERROR_CODE);                 
+    }
 
+    syslog(LOG_DEBUG,"Caught signal, exiting");
+    closelog();
+
+    if(-1 == remove(DATA_FILE))
+    {
+        perror("remove(DATA_FILE)");
+    }
 }
-	
-//Signal handler for Signals SIGTERM and SIGINT
-static void signal_handler(int signo)
+
+/*
+ * This functions cleans up exiting with -1
+ */
+void error_handler()
 {
-	close_err=close(client_sock_fd);
-	if(close_err == -1)
-	{
-		close_all();
-		perror("close client socket error");
-		exit(-1);
-	}
-
-	//Close server socket fd
-	close_err=close(serv_sock_fd);
-	if(close_err == -1)
-	{
-		close_all();
-		perror("close server socket error");
-		exit(-1);
-	}
-	
-	//Close regular file - output file fd
-	close_err=close(output_file_fd);
-	if(close_err == -1)
-	{
-		close_all();
-		perror("close output file error");
-		exit(-1);
-	}
-	//Once connection closed, logs status
-	syslog(LOG_DEBUG, "Closed connection from %s",IP_addr);
-	//After completing above procedure successfuly, exit logged
-	syslog(LOG_DEBUG,"Caught signal, exiting");
-	
-	//Close the syslog once all of logging is complete
-	closelog();
-	
-	//Delete and unlink the file
-	if(remove(TEST_FILE) == -1)
-	{
-		perror("file remove error");
-	}
+    close(main_sockfd);
+    close(client_sockfd);
+    close(data_file_fd);
+    closelog();
+    remove(DATA_FILE);
 }
+
 int main(int argc, char *argv[])
 {
-	//Opens a connection with facility as LOG_USER to Syslog.
+    //Opens a syslog with LOG_USER facility  
 	openlog("aesdsocket",0,LOG_USER);
-	
-	//Registering signal_handler as the handler for the signals SIGTERM 
-	//and SIGINT
-	//Reference: Ch 10 signals, Textbook: Linux System Programming
-	if (signal(SIGINT, signal_handler) == SIG_ERR) 
+    int extra_alloc = 1;
+    int initial_alloc_size = 500;
+    int present_location = 0;
+    int error_flag_getaddr = 0;
+    int bytes_read = 0;
+    int total_data_len = 0;
+    int yes = 1;
+    struct addrinfo hints;
+    struct addrinfo *res;
+    struct sockaddr_storage client_addr; // client's address information
+
+    /*
+     * Use the signal function to establish a signal handler for specific signals 
+     * Returns SIG_ERR on failure. On success - returns the previous signal handler (a function pointer)
+     */
+    if (SIG_ERR == signal(SIGINT, sig_handler)) 
 	{
-		fprintf (stderr, "Cannot handle SIGINT\n");
-		exit (EXIT_FAILURE);
-	}
-	
-	if (signal(SIGTERM, signal_handler) == SIG_ERR) 
-	{
-		fprintf (stderr, "Cannot handle SIGTERM\n");
-		exit (EXIT_FAILURE);
-	}
-	
-		//Adding only signals SIGINT and SIGTERM to an empty set to enable only them
-	sigset_t socket_set;
-	int rc = sigemptyset(&socket_set);
-	if(rc !=0)
-	{
-		perror("signal empty set error");
-		exit(-1);
-	}
-	rc = sigaddset(&socket_set,SIGINT);
-	if(rc !=0)
-	{
-		perror("error adding signal SIGINT to set");
-		exit(-1);
-	}
-	rc = sigaddset(&socket_set,SIGTERM);
-	if(rc !=0)
-	{
-		perror("error adding signal SIGTERM to set");
-		exit(-1);
-	}
-	
-	//With node as null and ai_flags as AI_PASSIVE, the socket address 
-	//will be suitable for binding a socket that will accept connections
-	struct addrinfo hints;
-	memset(&hints, 0, sizeof hints);
-	hints.ai_family = AF_UNSPEC;
-	hints.ai_socktype = SOCK_STREAM;
-	hints.ai_flags = AI_PASSIVE;
-	struct addrinfo *res;
-	
-	
-	//Node NULL and service set with port number as 9000, the pointer to
-	//linked list returned and stored in res
-	if (getaddrinfo(NULL, PORT , &hints, &res) != 0) 
-	{
-		perror("getaddrinfo error");
-		exit(-1);
+		fprintf (stderr, "signal() failed for SIGINT\n");
+		exit(ERROR_CODE);
 	}
 
-
-    
-    //Creating an end point for communication with type = SOCK_STREAM(connection oriented)
-	//and protocol =0 which allows to use appropriate protocol (TCP) here
-	serv_sock_fd=socket(res->ai_family,res->ai_socktype,res->ai_protocol);
-	if(serv_sock_fd==-1)
+	if (SIG_ERR == signal(SIGTERM, sig_handler)) 
 	{
-		perror("socket error");
-		exit(-1);
-	}
-	
-		
-	//Set options on socket to prevent binding errors from ocurring
-	int dummie =1;
-	if (setsockopt(serv_sock_fd, SOL_SOCKET, SO_REUSEADDR, &dummie, sizeof(int)) == -1) 
+		fprintf (stderr, "signal() failed for SIGTERM\n");
+		exit(ERROR_CODE);
+	} 
+
+    /*
+     * A variable sock_set of type sigset_t, which is a data structure used to represent a set of signals.
+     * sigemptyset : Used to initialize an empty signal set. It clears all signals from the set.
+     * sigaddset : Used to add specific signals (SIGINT and SIGTERM) to the signal set 
+     */
+    sigset_t sock_set;
+    if(sigemptyset(&sock_set) != 0)
+    {
+		perror("sigemptyset()");
+		exit(ERROR_CODE);        
+    }
+    if(sigaddset(&sock_set,SIGINT) != 0)
+    {
+		perror("sigaddset(): SIGINT ");
+		exit(ERROR_CODE);          
+    }
+    if(sigaddset(&sock_set,SIGTERM) != 0)
+    {
+		perror("sigaddset(): SIGTERM ");
+		exit(ERROR_CODE);          
+    }
+
+    //Create a server socket i.e., main_sockfd
+    main_sockfd = socket(AF_INET,SOCK_STREAM,0);
+    char ip_addr[INET6_ADDRSTRLEN];
+    if(-1 == main_sockfd)
+    {
+        perror("socket()");
+        exit(ERROR_CODE);
+    }
+
+    memset(&hints,0,sizeof(hints)); //make sure the struct is 0 first.
+    hints.ai_family = AF_INET; //Don't care IPv4 or IPv6
+    hints.ai_socktype = SOCK_STREAM; //TCP based socket
+    hints.ai_flags =  AI_PASSIVE; // fill in my IP for me
+    error_flag_getaddr = getaddrinfo(NULL,PORT,&hints,&res);
+
+    /*
+     *  Reference for gai_strerror : https://pubs.opengroup.org/onlinepubs/9699919799/functions/gai_strerror.html
+     */
+    if(error_flag_getaddr !=0)
+    {
+        fprintf(stderr, "getaddrinfo error: %s\n", gai_strerror(error_flag_getaddr));
+        exit(ERROR_CODE);
+    }
+
+    /*
+     * Reference for setsockopt : https://beej.us/guide/bgnet/html/#getaddrinfoprepare-to-launch
+     * To set options on a socket
+     */
+	if (-1 == setsockopt(main_sockfd, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(int))) 
 	{	
-		
-		perror("setsockopt error");
+		perror("setsockopt()");
+        exit(ERROR_CODE);
+    }
+
+    if(-1 == bind(main_sockfd, res->ai_addr, res->ai_addrlen)) 
+    {
+        perror("bind()");
+        freeaddrinfo(res);
+        close(main_sockfd);
+        exit(ERROR_CODE);
     }
     
-	//Assign address to the socket created
-	rc=bind(serv_sock_fd, res->ai_addr, res->ai_addrlen);
-	if(rc==-1)
-	{
-		perror("bind error");
-		exit(-1);
-	}
-	
-	//Below conditional check referenced from https://stackoverflow.com/questions/803776/help-comparing-an-argv-string
+     /* The requirement is that the program should fork after ensuring it can bind to port 9000, so doing it right after the bind step
+     * Reference : https://learning.oreilly.com/library/view/linux-system-programming/0596009585/ch05.html#daemons
+     * Credits : The exact code mentioned in the above book is used here to run in deamon mode
+     */
 	if(argc==2 && (!strcmp(argv[1], "-d")))
 	{
-		//Below code reference from chapter 5 of textbook reference
-		//Child forked to create a daemon custom named customd
-		pid_t customd;
-		customd = fork();
-		if (customd == -1)
+		pid_t pid;
+        /* create new process */
+		pid = fork();
+		if (-1 == pid)
 		{
-			perror("fork error");
-			exit(-1);
+			perror("fork()");
+			exit(ERROR_CODE);
 		}
-		else if (customd != 0 )
+		else if (pid != 0 )
 		{
-			//Exit parent, to make child get reparented to init_parent
-			exit (EXIT_SUCCESS);
-		}
-
-		//Create new session and process group
-		if(setsid() == -1) 
-		{
-			perror("setsid");
-			exit(-1);
-		}
-		
-
-		//Set working to root directory
-		if (chdir("/") == -1)
-		{
-			exit(-1);
-		}
-		
-		//Should close if any open files, skipping that helre
-		//Redirect Stdin,stdout,stderr to /dev/null
-		open ("/dev/null", O_RDWR); 
-		dup (0); 
-		dup (0); 
-	}
-	
-	//If bind passes, open a new file to store the packet that will be read
-	output_file_fd=open(TEST_FILE,O_CREAT|O_RDWR,0644);
-	if(output_file_fd == -1)
-	{
-		perror("error opening file at /var/temp/aesdsocketdata");
-		exit(-1);
-	}
-	
-	//Post binding, the server is running on the port 9000, so now
-	//remote connections (client) can listen to server
-	//Backlog - queue of incoming connections before being accepted to send/recv 
-	//backlog set as 4, can be reduced to a lower number since we get only 1 connection
-	if(listen(serv_sock_fd,BACKLOG)==-1)
-	{
-		perror("error listening");
-		exit(-1);
-	}
-	freeaddrinfo(res);
-	while(1)
-	{
-		
-		socklen_t conn_addr_len=sizeof(conn_addr);
-		//Accept an incoming connection
-		client_sock_fd = accept(serv_sock_fd, (struct sockaddr *)&conn_addr, &conn_addr_len);
-		if(client_sock_fd ==-1)
-		{
-			//If no incoming connection, graceful termination done
-			close_all();
-			exit(-1);
-		}
-		
-		//Below line of code reference: https://beej.us/guide/bgnet/html/
-		//Check if the connection address family is IPv4 or IPv6 and retrieve accordingly
-		//Convert the binary to text before logging 
-		inet_ntop(AF_INET, get_in_addr((struct sockaddr *)&conn_addr), IP_addr, sizeof(IP_addr));
-        syslog(LOG_DEBUG,"Accepted connection from %s", IP_addr);   
 			
-		//Allocates buffer and initializes value to 0, same as malloc
-		char *buf_data=calloc(CHUNK_SIZE,sizeof(char));
-		if(buf_data==NULL)
-		{
-			syslog(LOG_ERR,"Error: malloc failed");
-			close_all();
-			exit(-1);
+			exit(EXIT_SUCCESS);
 		}
-		int loc=0;
 
-		//block the signals before recieving packets and sending back
-		int merr=sigprocmask(SIG_BLOCK, &socket_set, NULL);
-		if(merr == -1)
+        /* The child process continues execution and performs the following steps to daemonize itself:*/
+		/* create new session and process group */
+        /* Detaching the process from the controlling terminal and creating a new session leader*/
+		if(-1 == setsid()) 
 		{
-			perror("sigprocmask block error");
-			close_all();
-			exit(-1);
+			perror("setsid()");
+			exit(ERROR_CODE);
+        }
+
+		/* set the working directory to the root directory */
+        /* This is done to ensure that the daemon process does not have any specific working directory tied to the terminal */
+		if (-1 == chdir("/"))
+		{
+			exit(ERROR_CODE);
 		}
-		//Recieve the data and store in updated location always if available
-		while((len=recv(client_sock_fd , buf_data + loc , CHUNK_SIZE , 0))>0)
-		{
-			if(len == -1)
+
+       
+        /* It redirects the standard input, output, and error to /dev/null by opening and duplicating file descriptors, 
+         * effectively detaching the process from the terminal and eliminating any input or output connections with the daemon.
+         */
+        /* redirect fd's 0,1,2 to /dev/null */
+		open ("/dev/null", O_RDWR);     /* stdin */
+        dup (0);                        /* stdout */
+        dup (0);                        /* stderror */
+	}    
+
+    //Create file to push data to /var/tmp/aesdsocketdata
+    data_file_fd = open(DATA_FILE,O_CREAT|O_RDWR,0777);
+	if(-1 == data_file_fd)
+	{
+	    perror("Error: Couldn't open the file at /var/temp/aesdsocketdata");
+		exit(ERROR_CODE);
+	}
+
+    if(-1 == listen(main_sockfd,BACKLOG)) //backlog assumed 10 i.e, >1
+    {
+        perror("listen()");
+        freeaddrinfo(res);
+        close(main_sockfd);
+        exit(ERROR_CODE);
+    }
+    freeaddrinfo(res); // all done with this structure
+    /*
+     *    Need a while(1) so that we can accept many client requests and handle them. 
+     */
+    while(1)
+    {
+        socklen_t client_len = sizeof(client_addr);
+        client_sockfd = accept(main_sockfd,(struct sockaddr*)&client_addr,&client_len);
+
+        if (-1 == client_sockfd) 
+        {
+            perror("accept()");
+            error_handler();
+            exit(ERROR_CODE);
+        }
+
+        /*
+         *  Reference for  inet_ntop : https://man7.org/linux/man-pages/man3/inet_ntop.3.html
+         */ 
+        if(NULL == inet_ntop(AF_INET, get_in_addr((struct sockaddr*)&client_addr),ip_addr, sizeof(ip_addr)))
+        {
+            perror("inet_ntop()");
+            error_handler();
+            exit(ERROR_CODE);
+        }
+        else
+        {
+            syslog(LOG_DEBUG,"Accepted connection from %s", ip_addr);
+        }
+
+        char *data_buf = (char*)malloc(initial_alloc_size*sizeof(char));
+        if(NULL == data_buf)
+        {
+            perror("malloc()");
+            error_handler();
+            exit(ERROR_CODE);
+        }
+
+        /*
+         * For completing any open connection operations i.e., resv() and send(), need to block the signal using sigprocmask()
+         * Blocking doesn't mean we miss the signal, the signal waits until its unblocked and gets serviced by kernel. 
+         */
+		if( -1 == sigprocmask(SIG_BLOCK, &sock_set, NULL))
+        {
+			perror("sigprocmask()");
+			error_handler();
+			exit(ERROR_CODE);
+        }
+
+        while((bytes_read = recv(client_sockfd,data_buf+present_location,initial_alloc_size,0)) > 0)
+        {
+
+            if(-1 == bytes_read)
+            {
+                perror("resv()");
+                error_handler();
+                exit(ERROR_CODE);          
+            }            
+            /*
+             * Reference to strchr function : https://man7.org/linux/man-pages/man3/strchr.3.html
+            */
+            if(strchr(data_buf ,'\n') != NULL)
+            {
+                break;
+            }
+				
+			//If new line is not received, need to keep incrementing the present location to avoid overwritting the previous data received
+			present_location+=bytes_read;
+            extra_alloc++;
+            data_buf = (char*)realloc(data_buf,(extra_alloc*initial_alloc_size)*sizeof(char));
+			if(NULL == data_buf)
 			{
-				perror("recv error");
-				close_all();
-				exit(-1);
+				syslog(LOG_ERR,"Error: realloc()");
+                free(data_buf);
+                error_handler();
+                exit(ERROR_CODE);
 			}
-			
-			if(strchr(buf_data ,'\n') != NULL)
-				break;
-			//Update the current location if newline not found
-			loc+=len;
-			
-			//if no new line character, need to add more memory and dynamically
-			//reallocate with an extra chunk
-			counter++;
-			buf_data=(char*)realloc(buf_data,((counter*CHUNK_SIZE)*sizeof(char)));
-			if(buf_data==NULL)
-			{
-				syslog(LOG_ERR,"Error: realloc failed");
-				close_all();
-				exit(-1);
-			}
-		}
+        }
+        if(-1 == bytes_read)
+        {
+            perror("resv()");
+            error_handler();
+            exit(ERROR_CODE);          
+        }
 
-			
-		//Write to file and position is updated
-		int werr = write(output_file_fd,buf_data,strlen(buf_data));
-		if (werr == -1)
-		{
-			perror("write error");
-			close_all();
-			exit(-1);
-		}
-		
+        //Write the data to /var/tmp/aesdsocketdata
+        if(-1 == write(data_file_fd,data_buf,strlen(data_buf)))
+        {
+			perror("write()");
+			error_handler();
+			exit(ERROR_CODE);           
+        }
 
-		
-		//Store last position of file
-		//int last_position = lseek(output_file_fd, 0, SEEK_CUR);
-		total_length += (strlen(buf_data));
-		//Place position ot beginnging
-		lseek(output_file_fd, 0, SEEK_SET);
-		char * send_data_buf=calloc(total_length,sizeof(char));
-		
-		
-		int rerr=read(output_file_fd,send_data_buf,total_length);
-		syslog(LOG_DEBUG,"%d",total_length);
-		if (rerr == -1)
-		{
-			perror("read error");
-			close_all();
-			exit(-1);
-		}
-		
-		syslog(LOG_DEBUG,"%s",send_data_buf);
-		//Send contents of buffer to client back
-		int serr=send(client_sock_fd,send_data_buf,total_length, 0);
-		if(serr == -1)
-		{
-			close_all();
-			perror("send error");
-			exit(-1);
-		}
+        //Need total data because we need to send back the whole data received till now
+        total_data_len += strlen(data_buf);
 
-		//Once complete, buffer cleared
-		free(buf_data);
-		free(send_data_buf);
-		
-		//Unblock the set of signals once complete
-		merr=sigprocmask(SIG_UNBLOCK, &socket_set, NULL);
-		if(merr == -1)
+        /* To make sure we sent the data from start of the file
+         * Move the cursor to start
+		 */
+        lseek(data_file_fd, 0, SEEK_SET);
+
+        //create a buffer to send data with total size available at /var/tmp/aesdsocketdata
+        char *send_to_client_buf = (char*)malloc(total_data_len*sizeof(char));
+        if(-1 == read(data_file_fd,send_to_client_buf,total_data_len))
+        {
+			perror("read()");
+			error_handler();
+			exit(ERROR_CODE);              
+        }
+
+        if(-1 == send(client_sockfd,send_to_client_buf,total_data_len, 0))
+        {
+			perror("send()");
+			error_handler();
+			exit(ERROR_CODE);              
+        }
+
+        free(send_to_client_buf);
+        free(data_buf);
+        syslog(LOG_DEBUG,"Closed connection from %s", ip_addr);
+
+		if(-1 == sigprocmask(SIG_UNBLOCK, &sock_set, NULL))
 		{
-			perror("sigprocmask unblock error");
-			close_all();
-			exit(-1);
+			perror("sigprocmask()");
+			error_handler();
+			exit(ERROR_CODE); 
 		}
     }
-	return 0;
-}	
+    return 0;
+}
+
